move to init region
